+++ dse_do_dashboard/visualization_pages/visualization_page.py
@@ -16,7 +16,9 @@ class VisualizationPage(ABC):
     * get_layout_children(self)
     """
     def __init__(self, dash_app, page_name:str='Default', page_id:str='default', url:str='default',
-                 input_table_names: List[str] = [], output_table_names: List[str] = []):
+                 input_table_names: List[str] = [], output_table_names: List[str] = [],
+                 enable_reference_scenario: bool = False,
+                 enable_multi_scenario: bool = False):
         """
         Best practice to specify the minimal set of input and output tables.
         Loading less tables speeds of the response time of the dashboard.
@@ -29,6 +31,8 @@ class VisualizationPage(ABC):
             url (str): url name
             input_table_names (List[str]): list of input table names required loading by the combined visualizations on this page. Or ['*'] for all input tables.
             output_table_names (List[str]): list of output table names required loading by the combined visualizations on this page. Or ['*'] for all output tables.
+            enable_reference_scenario (bool): if True and if reference_scenario_name not None, will load the reference scenario and add a ref_dm dataManager to the PlotlyManager
+            enable_multi_scenario (bool): if True and if there are more than 1 scenarios to compare in multi_scenario_names, will load multiple scenarios (including the currently selected) in ms_inputs and ms_outputs of the PlotlyManager
         TODO:
         * dash_app: DoDashApp. Due to circular imports, cannot specify type!
         """
@@ -39,12 +43,15 @@ class VisualizationPage(ABC):
 
         self.input_table_names = input_table_names  # Names of input tables required for any of the plots in this page
         self.output_table_names = output_table_names  # Names of output tables required for any of the plots in this page
+        self.enable_reference_scenario = enable_reference_scenario
+        self.enable_multi_scenario = enable_multi_scenario
         # if len(input_table_names) == 1 and input_table_names[0] == '*':
         #     self.input_table_names = dash_app.get_input_table_names()
         # else:
         #     self.input_table_names = input_table_names  # Names of input tables required for any of the plots in this page
 
-    def get_layout(self, scenario_name: str):
+    def get_layout(self, scenario_name: str,
+                   reference_scenario_name: str = None, multi_scenario_names: List[str] = None):
         """Returns layout of visualization page.
         TODO:
         * We are NOT using pattern-matching callback anynore, so the `id` of the Div is no long relevant
@@ -56,7 +63,11 @@ class VisualizationPage(ABC):
                     'type': 'tab_layout',
                     'index': self.page_id
                 },
-                children=self.get_layout_children(self.get_plotly_manager(scenario_name))
+                children=self.get_layout_children(self.get_plotly_manager(scenario_name,
+                                                                          reference_scenario_name=reference_scenario_name,
+                                                                          multi_scenario_names=multi_scenario_names,
+                                                                          enable_reference_scenario=self.enable_reference_scenario,
+                                                                          enable_multi_scenario=self.enable_multi_scenario))
             )
         ])
         return layout
@@ -117,8 +128,19 @@ class VisualizationPage(ABC):
             output_table_names = self.output_table_names
         return output_table_names
 
-    def get_plotly_manager(self, scenario_name: str):
-        return self.dash_app.get_plotly_manager(scenario_name, self.get_input_table_names(), self.get_output_table_names())
+    def get_plotly_manager(self, scenario_name: str, reference_scenario_name: str = None, multi_scenario_names: List[str] = None,
+                           enable_reference_scenario: bool = False,
+                           enable_multi_scenario: bool = False):
+        # reference_scenario_name = None
+        # multi_scenario_names = None
+        return self.dash_app.get_plotly_manager(scenario_name,
+                                                self.get_input_table_names(),
+                                                self.get_output_table_names(),
+                                                reference_scenario_name=reference_scenario_name,
+                                                multi_scenario_names=multi_scenario_names,
+                                                enable_reference_scenario=enable_reference_scenario,
+                                                enable_multi_scenario=enable_multi_scenario
+                                                )
 
     def set_dash_callbacks(self):
         """Define Dash callbacks for this page.